requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
anyio==3.5.0
async-timeout==3.0.1
attrs==22.2.0
beautifulsoup4==4.11.1
bing-image-downloader==1.1.2
blinker==1.4
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
cloudscraper==1.2.68
colorama==0.4.6
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
discord.py-self==1.9.2
discord.py-self-embed==1.0.6
et-xmlfile==1.1.0
fastapi==0.89.1
flit_core==3.6.0
fonttools==4.38.0
frozenlist==1.3.3
gingerit==0.9.0
google-search-results==2.4.1
h11==0.14.0
httptools==0.5.0
idna==3.4
kiwisolver==1.4.4
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.1
openai==0.26.2
openpyxl==3.1.0
packaging==23.0
Pillow==9.4.0
pip==22.3.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
selenium==3.141.0
setuptools==65.6.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.22.0
tqdm==4.64.1
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
XlsxWriter==3.0.7
yarl==1.8.2
Flask==2.1.0
gunicorn==20.1.0